separate handshakes into files
